server_new/requirements.txt
argon2-cffi==23.1.0
apscheduler==3.10.4
fastapi==0.115.6
starlette==0.41.3
uvicorn==0.34.0
docker==7.1.0
requests==2.32.3
PyJWT==2.10.1
pydantic==2.10.6
pydantic[email]
python-multipart==0.0.20
SQLAlchemy==2.0.38
alembic==1.14.1
Jinja2==3.1.4
psycopg2==2.9.10
websockets==15.0.1
